+++ packages/dom-snapshot/src/browser/processPage.js
@@ -89,9 +89,7 @@ function processPage(
     const canvasBlobs = buildCanvasBlobs(canvasElements);
     const frameDocs = extractFrames(docRoots);
 
-    const processFramesPromise = frameDocs.map(f =>
-      doProcessPage(f, f.defaultView.frameElement.src),
-    );
+    const processFramesPromise = frameDocs.map(f => doProcessPage(f));
     const processInlineFramesPromise = inlineFrames.map(({element, url}) =>
       doProcessPage(element.contentDocument, url),
     );